+++ test/tc_helpers.php
@@ -2,6 +2,7 @@
 require_once(__DIR__ . "/../src/app/helpers/modifier.img_url.php");
 require_once(__DIR__ . "/../src/app/helpers/modifier.img_width.php");
 require_once(__DIR__ . "/../src/app/helpers/modifier.img_height.php");
+require_once(__DIR__ . "/../src/app/helpers/modifier.img_color.php");
 
 class TcHelpers extends TcBase {
 
@@ -28,5 +29,27 @@ class TcHelpers extends TcBase {
 
 		$this->assertEquals(400,smarty_modifier_img_width($image_url,"400"));
 		$this->assertEquals(268,smarty_modifier_img_height($image_url,"400"));
+	}
+
+	function test_modifier_img_color(){
+		$image_url = "https://i.pupiq.net/i/65/65/27e/2927e/1272x920/9cUpr1_800x800xc_6c2a983e5ac4792b.jpg";
+		//
+		$this->assertEquals("#E7505B",smarty_modifier_img_color($image_url));
+		$this->assertEquals("#E7505B",smarty_modifier_img_color($image_url,"vibrant"));
+		$this->assertEquals("#E7505B",smarty_modifier_img_color($image_url,"vibrant or light_vibrant"));
+		$this->assertEquals("#D67E89",smarty_modifier_img_color($image_url,"light_vibrant"));
+		$this->assertEquals("#D67E89",smarty_modifier_img_color($image_url,"light_vibrant or vibrant"));
+		$this->assertEquals("#2CAC95",smarty_modifier_img_color($image_url,"dark_vibrant"));
+		$this->assertEquals("#6CB1AE",smarty_modifier_img_color($image_url,"muted"));
+		$this->assertEquals("#C69F9C",smarty_modifier_img_color($image_url,"light_muted"));
+		$this->assertEquals("#573031",smarty_modifier_img_color($image_url,"dark_muted"));
+		$this->assertEquals("",smarty_modifier_img_color($image_url,"nonsence"));
+
+		// This image has no light_vibrant color
+		$image_url = "https://i.pupiq.net/i/65/65/27c/2927c/1272x920/JuSG6C_800x800xc_648c5aa725d85c4b.jpg";
+		//
+		$this->assertEquals("#9E6A0B",smarty_modifier_img_color($image_url,"vibrant"));
+		$this->assertEquals("",smarty_modifier_img_color($image_url,"light_vibrant"));
+		$this->assertEquals("#9E6A0B",smarty_modifier_img_color($image_url,"light_vibrant or vibrant"));
 	}
 }