+++ test/tc_pupiq.php
@@ -10,10 +10,14 @@ class TcPupiq extends TcBase{
 		$this->assertEquals(800,$image->getWidth());
 		$this->assertEquals(537,$image->getHeight());
 
-		$this->assertEquals(123,$image->getUserId()); // !! TODO: Toto je podle PUPIQ_API_KEY a nikoli podle URL
+		$this->assertEquals(101,$image->getUserId()); // !! TODO: Toto je podle PUPIQ_API_KEY a nikoli podle URL
 		$this->assertEquals(30145,$image->getImageId()); // hexdec(75c1)
 
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.jpg","$image");
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_246234d94216d836.jpg","$image");
+
+		// svg image
+		$image = new Pupiq("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.svg");
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_246234d94216d836.svg","$image");
 
 		// no image url given
 		$image = new Pupiq();
@@ -21,7 +25,7 @@ class TcPupiq extends TcBase{
 
 		// hostname is automatically corrected
 		$image = new Pupiq("http://alpha.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.jpg");
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.jpg","$image");
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_246234d94216d836.jpg","$image");
 
 		// non pupiq image given
 		$image = new Pupiq();
@@ -37,26 +41,26 @@ class TcPupiq extends TcBase{
 	function test_formatting_image(){
 		$image = new Pupiq("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.jpg");
 
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x536_b1ae08e7e3612505.jpg",$image->getUrl("800x800"));
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x536_b1ae08e7e3612505.png",$image->getUrl("800x800,format=png"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x536_28caf7fe5e3a333b.jpg",$image->getUrl("800x800"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x536_28caf7fe5e3a333b.png",$image->getUrl("800x800,format=png"));
 
 		// cropping
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800xc_442ea0bf0c2cd3c9.jpg",$image->getUrl("800x800xcrop"));
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x400xc_ac756ea1866ff62f.jpg",$image->getUrl("800x400xcrop"));
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x400xct_6fcbe211b5f8fad3.jpg",$image->getUrl("800x400xcrop,top"));
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x400xcb_cb6aa1d62e976200.jpg",$image->getUrl("800x400xcrop,bottom"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800xc_d3d44c33f84756c0.jpg",$image->getUrl("800x800xcrop"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x400xc_a451a51e48dceb38.jpg",$image->getUrl("800x400xcrop"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x400xct_937564d9e86e192f.jpg",$image->getUrl("800x400xcrop,top"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x400xcb_d1a736c9bf38b8b1.jpg",$image->getUrl("800x400xcrop,bottom"));
 
 		// colored background
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800x112233_d6b32b5f565ccbd5.jpg",$image->getUrl("800x800x#112233"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800x112233_a3ea991200aa9d92.jpg",$image->getUrl("800x800x#112233"));
 
 
 		// transparent or colored background
 		// jpg -> colored border
 		$image = new Pupiq("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.jpg");
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800x002233_23ea47c40ef1c59a.jpg",$image->getUrl("800x800xtransparent_or_#002233"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800x002233_943d5488db5aa953.jpg",$image->getUrl("800x800xtransparent_or_#002233"));
 		// png -> transparent border
 		$image = new Pupiq("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x537_f6dd0e0882712ab1.png");
-		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800xt_e4493381151ea408.png",$image->getUrl("800x800xtransparent_or_#002233"));
+		$this->assertEquals("http://i.pupiq.net/i/2/2/75c/75c1/1300x872/gQs7Nv_800x800xt_e62352d612057e07.png",$image->getUrl("800x800xtransparent_or_#002233"));
 	}
 
 	function test_ToObject(){